moved app to App.jsx
