+++ package.json
@@ -9,7 +9,7 @@
   },
   "scripts": {
     "start": "webpack --mode development",
-    "develop": "webpack-dev-server",
+    "develop": "webpack serve --open",
     "build": "webpack --mode production",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
